+++ src/utils/searcher/similarity_searcher.py
@@ -12,8 +12,8 @@ class SimilaritySearcher:
 
         # Define argument types for the C functions
         self.lib.find_top_similar.argtypes = [
-            np.ctypeslib.ndpointer(dtype=np.float32),  # query
-            np.ctypeslib.ndpointer(dtype=np.float32),  # database
+            np.ctypeslib.ndpointer(dtype=np.float64),  # query
+            np.ctypeslib.ndpointer(dtype=np.float64),  # database
             ctypes.c_int,  # vector_dim
             ctypes.c_int,  # num_vectors
             np.ctypeslib.ndpointer(dtype=np.int32),  # top_indices
@@ -22,8 +22,8 @@ class SimilaritySearcher:
         ]
 
         self.lib.find_top_similar_batch.argtypes = [
-            np.ctypeslib.ndpointer(dtype=np.float32),  # queries
-            np.ctypeslib.ndpointer(dtype=np.float32),  # database
+            np.ctypeslib.ndpointer(dtype=np.float64),  # queries
+            np.ctypeslib.ndpointer(dtype=np.float64),  # database
             ctypes.c_int,  # vector_dim
             ctypes.c_int,  # num_vectors
             ctypes.c_int,  # num_queries
@@ -60,8 +60,8 @@ class SimilaritySearcher:
         assert query.ndim == 1 and database.ndim == 2, "Query and database must be 1D and 2D arrays respectively"
         assert query.shape[0] == database.shape[1], "Query and database must have the same vector dimension"
 
-        query = np.ascontiguousarray(query, dtype=np.float32)
-        database = np.ascontiguousarray(database, dtype=np.float32)
+        query = np.ascontiguousarray(query, dtype=np.float64)
+        database = np.ascontiguousarray(database, dtype=np.float64)
 
         vector_dim = query.shape[0]
         num_vectors = len(database)
@@ -100,8 +100,8 @@ class SimilaritySearcher:
         assert queries.ndim == 2 and database.ndim == 2, "Queries and database must be 2D arrays respectively"
         assert queries.shape[1] == database.shape[1], "Queries and database must have the same vector dimension"
 
-        queries = np.ascontiguousarray(queries, dtype=np.float32)
-        database = np.ascontiguousarray(database, dtype=np.float32)
+        queries = np.ascontiguousarray(queries, dtype=np.float64)
+        database = np.ascontiguousarray(database, dtype=np.float64)
 
         num_queries, vector_dim = queries.shape
         num_vectors = len(database)
@@ -128,8 +128,8 @@ if __name__ == "__main__":
 
     similarity_search = SimilaritySearcher(DLLPaths.SIMILARITY_SEARCH)
 
-    database = np.random.rand(2587, 1024).astype(np.float32)
-    query = np.random.rand(1024).astype(np.float32)
+    database = np.random.rand(2587, 1024).astype(np.float64)
+    query = np.random.rand(1024).astype(np.float64)
 
     start_time = time.time()
     for i in range(10000):
@@ -137,7 +137,7 @@ if __name__ == "__main__":
     end_time = time.time()
     print(f"Individual search time taken: {end_time - start_time} seconds")
 
-    queries = np.random.rand(10000, 1024).astype(np.float32)
+    queries = np.random.rand(10000, 1024).astype(np.float64)
     start_time = time.time()
     similarity_search.batch_search(queries, database)
     end_time = time.time()
